Don't list license multiple times in README
Maybe it should not be listed at all, but leaving it as is where the
authors file is currently mentioned.  Open for suggested improvements on
where to (not) list the license.  As discussed with nisse OOB, the one
place for the license is the license file, doesn't have to be in README.