Update trillian to v1.7.1 and grpc to v1.69.4
On the "go 1.22.0" directive in go.mod: trillian indirectly depends on
x/mod and x/tools, which both specify a dependency on "go@1.22.0".
Then using "go 1.22" breaks, because 1.22 is considered to old for the
1.22.0 dependencies. So both we and trillian have to include a third
".0" component, "go 1.22.0".

This is somewhat counter intuitive when all you're really interested
in is the language version; and different to the examples on
https://go.dev/ref/mod#go-mod-file-go, where none of the go directive
examples include a third component.

In later go versions (1.23 or 1.24?), the go tooling has been updated
to treat "go 1.x" more or less as an alias for "go 1.x.0". So at a
later update, we will likely be able to drop the ".0" and go back to
just two components, with "go 1.23" or "go 1.24".